User controller and more
The methods related to the user and the authentication middleware are separated from the code in the index.js file. The route files are created and the code in the main file is subsequently cleaned up. The code referring to the methods related to the creation and storage of video files is commented out. Further review is required to establish business rules for more complex functions such as sending notification emails and account activation.